requirements.txt
flask==3.0.3
gunicorn==22.0.0
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.2
google-cloud-speech==2.27.0
authlib==1.3.1
requests==2.32.3
anthropic==0.39.0
flask-cors==5.0.0
